Extract IDEA configuration
Also:
 * Simplify setting dependency on the launch task.
