+++ gradle-plugin/src/main/kotlin/io/spine/protodata/gradle/plugin/Plugin.kt
@@ -47,24 +47,16 @@ import io.spine.protodata.gradle.plugin.GeneratedSubdir.JAVA
 import io.spine.protodata.gradle.plugin.GeneratedSubdir.KOTLIN
 import io.spine.tools.code.manifest.Version
 import io.spine.tools.gradle.project.sourceSets
-import io.spine.tools.gradle.protobuf.generatedDir
-import io.spine.tools.gradle.protobuf.generatedSourceProtoDir
 import io.spine.tools.gradle.protobuf.protobufExtension
 import io.spine.util.theOnly
 import java.io.File
-import java.nio.file.Path
-import java.nio.file.Paths
 import org.gradle.api.Project
-import org.gradle.api.Task
 import org.gradle.api.file.SourceDirectorySet
 import org.gradle.api.tasks.Delete
 import org.gradle.api.tasks.SourceSet
 import org.gradle.kotlin.dsl.create
 import org.gradle.kotlin.dsl.exclude
 import org.gradle.kotlin.dsl.findByType
-import org.gradle.kotlin.dsl.getByType
-import org.gradle.plugins.ide.idea.model.IdeaModel
-import org.gradle.plugins.ide.idea.model.IdeaModule
 import org.gradle.api.Plugin as GradlePlugin
 
 /**
@@ -120,12 +112,6 @@ public class Plugin : GradlePlugin<Project> {
     }
 }
 
-private fun Project.createExtension(): Extension {
-    val extension = Extension(this)
-    extensions.add(CodegenSettings::class.java, EXTENSION_NAME, extension)
-    return extension
-}
-
 /**
  * Obtains an instance of the project [Extension] added by ProtoData Gradle Plugin.
  *
@@ -134,6 +120,18 @@ private fun Project.createExtension(): Extension {
 internal val Project.extension: Extension
     get() = extensions.findByType(CodegenSettings::class)?.run { this as Extension }
         ?: createExtension()
+
+private fun Project.createExtension(): Extension {
+    val extension = Extension(this)
+    extensions.add(CodegenSettings::class.java, EXTENSION_NAME, extension)
+    return extension
+}
+
+/**
+ * Obtains the name of the directory where ProtoData places generated files.
+ */
+internal val Project.targetBaseDir: String
+    get() = extension.targetBaseDir.toString()
 
 /**
  * Creates configurations for `protoDataRawArtifact` and user-defined classpath,
@@ -311,15 +309,6 @@ private fun Project.hasJavaOrKotlin(): Boolean {
 }
 
 /**
- * Obtains the name of the directory where ProtoData places generated files.
- */
-private val Project.targetBaseDir: String
-    get() {
-        val ext = extensions.getByType(CodegenSettings::class.java)
-        return ext.targetBaseDir.toString()
-    }
-
-/**
  * Obtains the `generated` directory for the given [sourceSet] and a language.
  *
  * If the language is not given, the returned directory is the root directory for the source set.
@@ -406,105 +395,12 @@ private fun GenerateProtoTask.generatedDir(language: String = ""): File =
  */
 private fun Project.handleLaunchTaskDependency(generateProto: GenerateProtoTask) {
     val sourceSet = generateProto.sourceSet
-    var launchTask: Task? = LaunchTask.find(project, sourceSet)
-    if (launchTask != null) {
-        launchTask.dependsOn(generateProto)
-    } else {
-        project.afterEvaluate {
-            val settingsTask =
-                project.tasks.withType(CreateSettingsDirectory::class.java).theOnly()
-            launchTask = createLaunchTask(settingsTask, sourceSet)
-            launchTask!!.dependsOn(generateProto)
+    LaunchTask.find(this, sourceSet)
+        ?.dependsOn(generateProto)
+        ?: afterEvaluate {
+            val settingsTask = tasks.withType(CreateSettingsDirectory::class.java).theOnly()
+            val launchTask = createLaunchTask(settingsTask, sourceSet)
+            launchTask.dependsOn(generateProto)
             createCleanTask(sourceSet)
         }
-    }
 }
-
-/**
- * Ensures that the sources generated by Protobuf Gradle plugin are
- * not included in the IDEA project.
- *
- * IDEA should only see the sources generated by ProtoData as
- * we define in [GenerateProtoTask.configureSourceSetDirs].
- */
-private fun Project.configureIdea() {
-    val thisProject = this
-    gradle.afterProject {
-        if (it == thisProject) {
-            pluginManager.withPlugin("idea") {
-                val idea = extensions.getByType<IdeaModel>()
-                idea.module.setupDirectories(thisProject)
-            }
-        }
-    }
-}
-
-private fun IdeaModule.setupDirectories(project: Project,) {
-
-    fun filterSources(sources: Set<File>, excludeDir: File): Set<File> =
-        sources.filter { !it.residesIn(excludeDir) }.toSet()
-
-    val protocOutput = project.file(project.generatedSourceProtoDir)
-    val protocTargets = project.protocTargets()
-    excludeWithNested(protocOutput.toPath(), protocTargets)
-    sourceDirs = filterSources(sourceDirs, protocOutput)
-    testSources.filter { !it.residesIn(protocOutput) }
-    generatedSourceDirs = project.generatedDir.resolve(protocTargets)
-        .map { it.toFile() }
-        .toSet()
-}
-
-/**
- * Obtains the path of the `generated` directory under the project root directory.
- */
-private val Project.generatedDir: Path
-    get() = projectDir.resolve(targetBaseDir).toPath()
-
-/**
- * Lists target directories for Protobuf code generation.
- *
- * The directory names are in the following format:
- *
- * `<source-set-name>/<builtIn-or-plugin-name>`
- */
-private fun Project.protocTargets(): List<Path> {
-    val protobufTasks = tasks.withType(GenerateProtoTask::class.java)
-    val codegenTargets = sequence {
-        protobufTasks.forEach { task ->
-            val sourceSet = task.sourceSet.name
-            val builtins = task.builtins.map { builtin -> builtin.name }
-            val plugins = task.plugins.map { plugin -> plugin.name }
-            val combined = builtins + plugins
-            combined.forEach { subdir ->
-                yield(Paths.get(sourceSet, subdir))
-            }
-        }
-    }
-    return codegenTargets.toList()
-}
-
-/**
- * Excludes the given directory and its subdirectories from
- * being seen as ones with the source code.
- *
- * The primary use of this extension is to exclude `build/generated/source/proto` and its
- * subdirectories to avoid duplication of types in the generated code with those in
- * produced by ProtoData under the `$projectDir/generated/` directory.
- */
-private fun IdeaModule.excludeWithNested(directory: Path, subdirs: Iterable<Path>) {
-    excludeDirs.add(directory.toFile())
-    directory.resolve(subdirs).forEach {
-        excludeDirs.add(it.toFile())
-    }
-}
-
-private fun Path.resolve(subdirs: Iterable<Path>): List<Path> =
-    subdirs.map {
-        resolve(it)
-    }
-
-/**
- * Tells if this file resides in the given [directory].
- */
-private fun File.residesIn(directory: File): Boolean =
-    canonicalFile.startsWith(directory.absolutePath)
